# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "媒体管理器 API 文档，包括媒体库管理、媒体文件管理等功能",
        "title": "媒体管理器 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:23347",
    "basePath": "/",
    "paths": {
...
        "/api/resources": {
            "get": {
                "description": "获取系统中所有资源文件的列表，支持分页和类型筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资源管理"
                ],
                "summary": "获取资源文件列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量，默认10，最大100",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "资源类型筛选：image | video | audio | document",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "媒体库ID筛选",
                        "name": "library_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.ResourceListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/resources/upload": {
            "post": {
                "description": "上传各类资源文件（图片、文档、音频、视频等）到系统",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资源管理"
                ],
                "summary": "上传资源文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "资源文件（最大10MB）",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "媒体库ID，可选",
                        "name": "library_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "资源类型（系统自动识别）",
                        "name": "resource_type",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.UploadResourceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "文件上传失败或文件过大",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "保存失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/resources/{id}": {
            "get": {
                "description": "根据ID获取资源文件，返回文件内容或重定向到CDN地址",
                "produces": [
                    "application/octet-stream",
                    "image/*",
                    "text/*"
                ],
                "tags": [
                    "资源管理"
                ],
                "summary": "获取资源文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源ID（示例: 1）",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "资源文件内容",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "302": {
                        "description": "CDN重定向",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "无效的资源ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "资源不存在或内容不可用",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "永久删除指定的资源文件，包括本地和CDN存储",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "资源管理"
                ],
                "summary": "删除资源文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源ID（示例: 1）",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "无效的资源ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "资源不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ResourceFile": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer",
                    "example": 1640995200
                },
                "file_name": {
                    "type": "string",
                    "example": "example.jpg"
                },
                "file_size": {
                    "type": "integer",
                    "example": 1048576
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "mime_type": {
                    "type": "string",
                    "example": "image/jpeg"
                },
                "storage_type": {
                    "type": "string",
                    "example": "cdn"
                },
                "type": {
                    "type": "string",
                    "example": "image"
                },
                "url": {
                    "type": "string",
                    "example": "https://cdn.example.com/files/example.jpg"
                }
            }
        },
        "api.ResourceListResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "resources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ResourceFile"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "api.UploadResourceResponse": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string",
                    "example": "example.jpg"
                },
                "file_size": {
                    "type": "integer",
                    "example": 1048576
                },
                "mime_type": {
                    "type": "string",
                    "example": "image/jpeg"
                },
                "resource_id": {
                    "type": "integer",
                    "example": 1
                },
                "type": {
                    "type": "string",
                    "example": "image"
                },
                "url": {
                    "type": "string",
                    "example": "https://cdn.example.com/files/example.jpg"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.ResponseCode"
                        }
                    ]
                },
                "data": {
                    "description": "数据"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                },
                "timestamp": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
        "response.ResponseCode": {
            "type": "integer",
            "enum": [
                200,
                400,
                401,
                403,
                404,
                405,
                409,
                422,
                429,
                500,
                501,
                502,
                503
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeMethodNotAllowed",
                "CodeConflict",
                "CodeUnprocessableEntity",
                "CodeTooManyRequests",
                "CodeInternalServerError",
                "CodeNotImplemented",
                "CodeBadGateway",
                "CodeServiceUnavailable"
            ]
        }
    },
...
}